Server/requirements.txt
ultralytics
flask
flask-cors
pytesseract # remove this since it is not used
tensorflow
ocr-nanonets-wrapper